+++ Addons/Backpacks/hpp/XtdGearInfos.hpp
@@ -1,19 +1,15 @@
 class XtdGearInfos
 {
-    class CfgVehicles
-    {
-        class ONI_Recon_Rucksack
-        {
+    class CfgVehicles {
+        class ONI_Recon_Rucksack {
             model = "ONI_Recon_Backpack";
             variant = "rifleman";
         };
-        class ONI_Recon_Rucksack_Medic
-        {
+        class ONI_Recon_Rucksack_Medic {
             model = "ONI_Recon_Backpack";
             variant = "medic";
         };
-        class ONI_Recon_Rucksack_IC
-        {
+        class ONI_Recon_Rucksack_IC {
             model = "ONI_Recon_Backpack";
             variant = "IC";
         };
